+++ lib/src/dialog_helper.dart
@@ -3,6 +3,7 @@ import 'dart:ui';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+import 'package:overlay_dialog/overlay_dialog.dart';
 import 'package:overlay_dialog/src/animation/appear_widget.dart';
 import 'package:overlay_dialog/src/dialog_widget.dart';
 
@@ -14,34 +15,55 @@ class DialogHelper {
 
   static final DialogHelper _instance = DialogHelper._();
 
+  static DialogStyle defaultStyle = DialogStyle.adaptive;
+
   factory DialogHelper() => _instance;
 
   DialogHelper._();
 
-  OverlayEntry _currentOverlay;
-  Function _currentCallback;
-  StreamController<double> _currentController;
+  List<IndexedData<OverlayEntry>> _currentOverlay = [];
+  List<IndexedData<Future<bool> Function()>> _currentCallback = [];
+  List<IndexedData<StreamController<double>>> _currentController = [];
+
+  Future<bool> onWillPop() {
+    return _currentCallback?.last?.data?.call() ?? Future.value(true);
+  }
 
   // Shows the dialog
-  void show(BuildContext context, DialogWidget dialog, {bool rootOverlay = true}) {
-    hideImmediate(context);
+  void show(BuildContext context, DialogWidget dialog, {bool rootOverlay = true, int id, bool hidePrevious = true}) {
+    if (hidePrevious || id == null) hideImmediate(context);
 
     OverlayState overlayState = rootOverlay
       ? context.findRootAncestorStateOfType<OverlayState>()
       : context.findAncestorStateOfType<OverlayState>();
 
     if (dialog.closable) {
-      _currentCallback = () {
-        hide(context);
-        return Future.value(false);
-      };
+      _currentCallback.add(
+        IndexedData<Future<bool> Function()>(
+          id: id,
+          data: () {
+            hide(context, id: id);
+            return Future.value(false);
+          })
+      );
 
-      ModalRoute.of(context).addScopedWillPopCallback(_currentCallback);
+    } else {
+      _currentCallback.add(
+        IndexedData<Future<bool> Function()>(
+          id: id,
+          data: () => Future.value(true)
+        )
+      );
     }
 
-    _currentController = StreamController();
+    if (_currentCallback.length == 1) {
+      print("add callback");
+      ModalRoute.of(context)?.addScopedWillPopCallback(onWillPop);
+    }
 
-    _currentOverlay = OverlayEntry(
+    final StreamController<double> controller = StreamController();
+
+    final OverlayEntry overlayEntry = OverlayEntry(
       builder: (_) => AppearWidget(
         child: Stack(
           children: <Widget>[
@@ -51,53 +73,91 @@ class DialogHelper {
                 color: BACKGROUND_COLOR,
               ),
             ),
-            dialog,
+            dialog
           ],
         ),
-        progress: _currentController.stream,
+        progress: controller.stream,
         duration: DEFAULT_DURATION,
         style: AppearStyle.opacity,
       ),
     );
 
-    overlayState.insert(_currentOverlay);
-    _currentController.add(1.0);
+    _currentController.add(
+      IndexedData<StreamController<double>>(
+        id: id,
+        data: controller
+      )
+    );
+
+    _currentOverlay.add(
+      IndexedData<OverlayEntry>(
+        id: id,
+        data: overlayEntry
+      )
+    );
+
+    overlayState.insert(overlayEntry);
+    controller.add(1.0);
   }
 
   // Hide opened dialog with animation
-  void hide(BuildContext context) {
-    if (_currentCallback != null) {
-      ModalRoute.of(context)?.removeScopedWillPopCallback(_currentCallback);
-      _currentCallback = null;
+  void hide(BuildContext context, {int id}) {
+    if (_currentCallback.length == 1 || id == null) {
+      ModalRoute.of(context)?.removeScopedWillPopCallback(onWillPop);
     }
 
-    if (_currentController != null) {
-      _currentController.add(0.0);
+    _currentController
+      .where((controller) => controller.id == id || id == null)
+      .forEach((controller) => controller.data.add(0.0));
 
-      Future
-        .delayed(DEFAULT_DURATION)
-        .then((_) {
-          _currentController?.close();
-          _currentController = null;
-          _currentOverlay?.remove();
-          _currentOverlay = null;
-        })
-        .catchError((error){});
-    }
+    Future
+      .delayed(DEFAULT_DURATION)
+      .then((_) => _hide(id))
+      .catchError((error){});
   }
 
   // Hide opened dialog without animation, clear closable callback if any
-  void hideImmediate(BuildContext context) {
+  void hideImmediate(BuildContext context, {int id}) {
     //reset back press handler
-    if (_currentCallback != null) {
-      ModalRoute.of(context)?.removeScopedWillPopCallback(_currentCallback);
-      _currentCallback = null;
+    if (_currentCallback.length == 1 || id == null) {
+      ModalRoute.of(context)?.removeScopedWillPopCallback(onWillPop);
     }
 
-    _currentController?.close();
-    _currentController = null;
-
-    _currentOverlay?.remove();
-    _currentOverlay = null;
+    _hide(id);
   }
+
+  void _hide(int id) {
+    _currentOverlay.removeWhere((overlay) {
+      if (overlay.id == id || id == null) {
+        try {
+          overlay.data.remove();
+        } catch(error) {}
+
+        return true;
+      }
+
+      return false;
+    });
+
+    _currentController.removeWhere((controller) {
+      if (controller.id == id || id == null) {
+        try {
+          controller.data.close();
+        } catch(error) {}
+
+        return true;
+      }
+
+      return false;
+    });
+
+    _currentCallback.removeWhere((callback) => callback.id == id || id == null);
+  }
+}
+
+class IndexedData<T> {
+  int id;
+  T data;
+
+  IndexedData({this.id, this.data});
 }