+++ lib/src/dialog_widget.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/widgets.dart';
+import 'package:overlay_dialog/overlay_dialog.dart';
 import 'package:overlay_dialog/src/dialog_factory.dart';
 import 'package:overlay_dialog/src/entity/dialog_action.dart';
 import 'package:overlay_dialog/src/entity/dialog_style.dart';
@@ -12,24 +13,25 @@ class DialogWidget extends StatelessWidget {
   final Widget _widget;
 
   DialogWidget.alert({
-    DialogStyle style = DialogStyle.adaptive,
+    DialogStyle style,
     String title,
     String content,
     List<DialogAction> actions,
     this.closable = true,
-  }): _widget = DialogFactory(style).alert(title, content, actions);
+  }): _widget = DialogFactory(style ?? DialogHelper.defaultStyle).alert(title, content, actions);
 
   /*DialogWidget.input();*/
 
   DialogWidget.progress({
-    DialogStyle style = DialogStyle.adaptive,
+    DialogStyle style,
     this.closable = false
-  }): _widget = DialogFactory(style).progress();
+  }): _widget = DialogFactory(style ?? DialogHelper.defaultStyle).progress();
 
   DialogWidget.custom({
+    DialogStyle style,
     @required Widget child,
     this.closable = true
-  }): _widget = Center(child: child);
+  }): _widget = DialogFactory(style ?? DialogHelper.defaultStyle).custom(child);
 
   @override
   Widget build(BuildContext context) {